11.10

+++ Src/Sxc/ToSic.Sxc/Apps/Parts/AppsManager.cs
@@ -1,46 +1,45 @@
 ﻿using System;
 using System.IO;
-using ToSic.Eav;
 using ToSic.Eav.Apps;
 using ToSic.Eav.Apps.Parts;
-using ToSic.Eav.Logging;
+using ToSic.Eav.Plumbing;
 
 namespace ToSic.Sxc.Apps
 {
-    public class AppsManager: ZonePartRuntimeBase
+    public class AppsManager: ZonePartRuntimeBase<ZoneRuntime, AppsManager>
     {
-        internal CmsZones CmsZones;
+        #region Constructor / DI
 
-        internal AppsManager(CmsZones cmsZones, ILog parentLog) : base(cmsZones, parentLog, "Cms.AppsRt")
+        public AppsManager(Lazy<ZoneManager> zoneManagerLazy, IServiceProvider serviceProvider) : base(serviceProvider, "Cms.AppsRt")
         {
-            CmsZones = cmsZones;
+            _zoneManagerLazy = zoneManagerLazy;
         }
+        private readonly Lazy<ZoneManager> _zoneManagerLazy;
+
+        #endregion
 
 
-
-        internal void RemoveAppInTenantAndEav(int appId)
+        internal void RemoveAppInSiteAndEav(int appId, bool fullDelete)
         {
             var zoneId = ZoneRuntime.ZoneId;
             // check portal assignment and that it's not the default app
-            //if (zoneId != CmsZones.Env.ZoneMapper.GetZoneId(tenant.Id))
-            //    throw new Exception("This app does not belong to portal " + tenant.Id);
-
             if (appId == ZoneRuntime.DefaultAppId)
                 throw new Exception("The default app of a zone cannot be removed.");
 
             // todo: maybe verify the app is of this portal; I assume delete will fail anyhow otherwise
 
             // Prepare to Delete folder in dnn - this must be done, before deleting the app in the DB
-            var sexyApp = Factory.Resolve<App>().InitNoData(new AppIdentity(zoneId, appId), null);
+            var sexyApp = ServiceProvider.Build<App>().InitNoData(new AppIdentity(zoneId, appId), null);
             var folder = sexyApp.Folder;
             var physPath = sexyApp.PhysicalPath;
 
             // now remove from DB. This sometimes fails, so we do this before trying to clean the files
             // as the db part should be in a transaction, and if it fails, everything should stay as is
-            new ZoneManager(zoneId, Log).DeleteApp(appId);
+            _zoneManagerLazy.Value.Init(zoneId, Log).DeleteApp(appId, fullDelete);
 
             // now really delete the files - if the DB didn't end up throwing an error
-            if (!string.IsNullOrEmpty(folder) && Directory.Exists(physPath))
+            // ...but only if it's a full-delete
+            if (fullDelete && !string.IsNullOrEmpty(folder) && Directory.Exists(physPath))
                 Directory.Delete(physPath, true);
 
         }
